requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.5
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-json-logger==2.0.7

# Admin Panel
django>=5.0
django-jazzmin>=3.0.0
django-environ>=0.11.2
